Add files via upload
he 19th Hole sells golf clubs at its golf outlet stores throughout the US.  The company knows that the demand for its club varies considerably with price.  In the last year, the demand at each price level was recorded in the table below. The Table can be found in the Excel spreadsheet.
The company wants to estimate the relationship between demand and price and then use this estimated relationship to answer the following questions: 

Assuming the unit cost of producing a set of clubs is $250, and the price must be a multiple of $10, what price should the company charge to maximize its profit?

How does the optimal price depend on the unit cost of producing a set of clubs?

Is the model an accurate representation of reality?