Add files via upload
Model Set-Up
Create a new column called Year Number. 2017 should go with Year Number 1, 2018 should go with Year Number 2, and so on. 

Create a scatter plot of the annual revenue. The Year Number data should go with the 
�
x-axis and the Revenue data should go with the 
�
y-axis. Make sure that your chart has an accurate title and axis labels.

Fit a power function and an exponential function to these data. Display the equations for the two trendlines on the chart.

Note: The Year Number column is needed to make the data easier to model. If you tried to make a scatterplot of the years and the revenue, you would get coefficients of 0 for the exponential and power trendlines. This is because plugging large numbers like 2017 into exponential or power function results in numbers that are very large. If your exponential and/or power functions are not showing up on your chart or have coefficients of 0, ensure that you have set up your Year Number column correctly and used that in your scatterplot instead of the original years.
After you have found an exponential model and power function model, set up a table to find the mean absolute percentage error (MAPE) for each model.  
Use the model that fits better to predict Peloton’s revenue in 2023, in millions of dollars.